security integration

+++ src/main/java/edu/ban7/rh_cesi_26/controller/ResourceController.java
@@ -3,11 +3,15 @@ package edu.ban7.rh_cesi_26.controller;
 import com.fasterxml.jackson.annotation.JsonView;
 import edu.ban7.rh_cesi_26.dao.ResourceDao;
 import edu.ban7.rh_cesi_26.model.Resource;
+import edu.ban7.rh_cesi_26.security.AppUserDetails;
+import edu.ban7.rh_cesi_26.security.IsAdmin;
+import edu.ban7.rh_cesi_26.security.IsUser;
 import edu.ban7.rh_cesi_26.view.ResourceView;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.access.prepost.PreAuthorize;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -15,6 +19,7 @@ import java.util.Optional;
 
 @RestController
 @RequestMapping("/api/resource")
+@IsUser
 public class ResourceController {
 
     @Autowired
@@ -22,7 +27,6 @@ public class ResourceController {
 
     @GetMapping("/list")
     @JsonView(ResourceView.class)
-    @PreAuthorize("hasAnyRole('ROLE_ADMIN')")
     public List<Resource> getResources() {
         return resourceDao.findAll();
     }
@@ -43,20 +47,31 @@ public class ResourceController {
 
     @PostMapping
     @JsonView(ResourceView.class)
-    public ResponseEntity<Resource> create(@RequestBody Resource resource) {
-        resourceDao.save(resource);
+    public ResponseEntity<Resource> create(
+            @RequestBody Resource resource,
+            @AuthenticationPrincipal AppUserDetails appUserDetails) {
 
+        resource.setOwner(appUserDetails.getUser());
+        resourceDao.save(resource);
         return new ResponseEntity<>(resource,HttpStatus.CREATED);
     }
 
     @DeleteMapping("/{id}")
-    public ResponseEntity<Void> delete(@PathVariable int id) {
-
+    public ResponseEntity<Void> delete(
+            @PathVariable int id,
+            @AuthenticationPrincipal AppUserDetails appUserDetails
+    ) {
         Optional<Resource> optionalResource = resourceDao.findById(id);
 
         if(optionalResource.isEmpty()) {
             //return ResponseEntity.notFound().build();
             return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
+
+        //si l'utilisateur n'est ni admin ni propriétaire de la resource
+        if(!appUserDetails.getUser().isAdmin()
+                && appUserDetails.getUser().getId() != optionalResource.get().getOwner().getId()) {
+            return new ResponseEntity<>(HttpStatus.FORBIDDEN);
         }
 
         resourceDao.deleteById(id);
